+++ lib/src/color_wheel_picker.dart
@@ -9,7 +9,7 @@ import 'package:flutter/material.dart';
 // The handy part is that if it gets in the way in debugging, it is an easy
 // toggle to turn it off there too. Often I just leave them true if it is one
 // I want to see in dev mode, unless it is too chatty.
-const bool _debug = !kReleaseMode && true;
+const bool _debug = !kReleaseMode && false;
 
 /// A HSV color wheel based color picker for Flutter, used by FlexColorPicker.
 ///